requirement.txt
numpy
pycuda
nvidia-tensorrt==8.2.1.8
opencv-python
onnxruntime-gpu

